feat(dto): changed code to use newly added dto classes

+++ src/main/java/mfandrade/sampleapp/controllers/ClientController.java
@@ -1,7 +1,8 @@
 package mfandrade.sampleapp.controllers;
 
-import java.time.Instant;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -13,6 +14,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.transaction.Transactional;
+import mfandrade.sampleapp.controllers.dto.ClientDTO;
+import mfandrade.sampleapp.controllers.dto.ClientFormDTO;
+import mfandrade.sampleapp.controllers.dto.ClientUpdateFormDTO;
 import mfandrade.sampleapp.entities.Client;
 import mfandrade.sampleapp.repositories.ClientRepository;
 
@@ -23,31 +27,39 @@ public class ClientController {
   @Autowired
   private ClientRepository clientRepository;
 
-  public List<Client> list() {
+  public List<ClientDTO> list() {
     List<Client> clients = clientRepository.findAll();
-    return clients;
+    return ClientDTO.convert(clients);
   }
 
   @Transactional
   @PostMapping
-  public void save(@RequestBody Client data) {
-    Instant now = Instant.now();
-    data.setCreatedAt(now);
-    data.setUpdatedAt(now);
-    clientRepository.save(data);
+  public ClientDTO save(@RequestBody ClientFormDTO form) {
+    Client client = form.convert();
+    clientRepository.save(client);
+    return new ClientDTO(client);
   }
 
   @Transactional
-  @PutMapping
-  public void update(@RequestBody Client data) {
-    data.setUpdatedAt(Instant.now());
-    clientRepository.save(data);
+  @PutMapping("/{cod}")
+  public ClientDTO update(@PathVariable Long cod, @RequestBody ClientUpdateFormDTO form) {
+    final Optional<Client> registry = clientRepository.findById(cod);
+    if (!registry.isPresent()) {
+      throw new NoSuchElementException(String.format("Client #%s not found", cod));
+    }
+    Client client = registry.get();
+    form.update(client);
+
+    return new ClientDTO(client);
   }
 
   @Transactional
   @DeleteMapping("/{cod}")
   public void delete(@PathVariable Long cod) {
-    clientRepository.deleteById(cod);
+    final Optional<Client> registry = clientRepository.findById(cod);
+    if (registry.isPresent()) {
+      clientRepository.deleteById(cod);
+    }
   }
 
 }
